requirements.txt
fastapi==0.95.0
requests==2.28.2
typing_extensions==4.5.0
airspeed
watchdog
streamlit_option_menu
streamlit
gpt4all-j
GitPython
